Add docker pull functionality to the Makefile.
This lets a user add their own list of docker base images into the
vagrant vm base image without needing to rewrite code. Simply make the
changes in the versions/*.sh file that you're using to build.

XXX: this doesn't work yet, but when it does, I can amend to remove this
line and add in this one:
Special thanks to rwmjones, for constantly having the magic solution.